+++ pyincore/restorationservice.py
@@ -9,6 +9,7 @@ from urllib.parse import urljoin
 
 from pyincore import IncoreClient
 from pyincore.dfr3service import Dfr3Service
+from pyincore.utils import return_http_response
 
 
 class RestorationService(Dfr3Service):
@@ -63,4 +64,4 @@ class RestorationService(Dfr3Service):
             payload['space'] = space
 
         r = self.client.get(url, params=payload, timeout=timeout, **kwargs)
-        return self.return_http_response(r).json()
+        return return_http_response(r).json()
